requirements.txt
torch
torchvision
timm
streamlit
pillow
requests
google-generativeai
python-dotenv